renames components. moves redux and apollo to container. updates names. adds loading state bool to search for later.
