+++ src/state/forms/reducer.js
@@ -1,2 +1,2 @@
 import formReducer from 'abyss-form/lib/reducer';
-export default formReducer(['login', 'verify', 'profile', 'storySearch', 'publish']);
+export default formReducer(['login', 'verify', 'profile', 'search', 'publish']);
